Enable some more warnings in 2022 Makefile, and fix some of them

+++ 2022/Cpp/src/Day8.cpp
@@ -46,7 +46,7 @@ std::pair<std::string, std::string> day8combined(std::ifstream input) {
 			}
 
 			hidden = false;
-			for (int64_t x1 = x + 1; x1 < map[y].size(); x1++) {
+			for (size_t x1 = x + 1; x1 < map[y].size(); x1++) {
 				if (map[y][x1] >= c) {
 					hidden = true;
 					scenic *= x1 - x;
@@ -74,7 +74,7 @@ std::pair<std::string, std::string> day8combined(std::ifstream input) {
 			}
 
 			hidden = false;
-			for (int64_t y1 = y + 1; y1 < map.size(); y1++) {
+			for (size_t y1 = y + 1; y1 < map.size(); y1++) {
 				if (map[y1][x] >= c) {
 					hidden = true;
 					scenic *= y1 - y;
